adding somewhat generic impl of inference tester session runner (aka trainer.eval)
