+++ conda-env.yml
@@ -11,7 +11,8 @@ dependencies:
   - matplotlib
   - numpy
   - opencv
-  - pillow
+  # PILLOW_VERSION removed which causes torchvision<=0.4.0 to fail
+  - pillow<7
   - pip
   - pytorch<1.3
   - pyyaml
